moved token fetching to app.jsx, made some routes inaccessible when not logged in
